matmul: Add cuda kernel for tiled matrix multiply
The algorithm is roughly the same as for naive implementation, but the
memory access is done differently, by reading tiles into shared memory
first and then doing the computation.

This improves the performance for 1024x1024 matrices by ~26%
and for 2048x2048 matrices by ~31%.

The accesses might probably still be done even better by reading more
tiles at once. Something to explore for future.