matmul: Add tests/files/*.safetensors to .gitignore
We don't store safetensors files in the repo, because we easily reach
storage limits. Instead, user is expected to regenerate the files using
the tests/gen-safetensors.py script.

I have spent too much time trying to make it work. LFS works fine with
pribately hosted git repo and lfs storage (server), but makes it hard to
publish the code in the open.